toolbox/data_toolbox/requirements.txt
# Data Toolbox Requirements
# 다른 팀에서 data_toolbox를 MCP 리소스로 사용하기 위한 최소 요구사항

# Database - PostgreSQL
psycopg-pool>=3.0.0
psycopg[binary]>=3.0.0

# Vector Database - ChromaDB
chromadb>=1.0.0

# Data Processing
pandas>=2.0.0

# HTTP Client
requests>=2.31.0

# Date/Time Utilities
python-dateutil>=2.8.0

# FastMCP Server
fastmcp==2.11.3
